chore: refines shimmer, adds to list view and CodeEditor element
